Eliminar CsvHelper y simplificar lectura de CSV
